Move static files to static directory
