+++ static/utils.ts
